[L4TLauncher] Disable rootfs validation
This patch is copied from external/sota/recipes-bsp/uefi/
edk2-firmware-tegra/0003-l4tlauncher-disable-a-b-rootfs-validation.patch.

From the patch summary:
We don't support a/b rootfs, so disable the runtime bootcount and slot
validation to avoid requiring nvbootctrl to be called during boot (to
clear up the variables).

This is to avoid forcing the device to boot in recovery after 3 boots
without confirmation from nvbootctrl.

Signed-off-by: Tushar Pankaj <tushar.s.pankaj@gmail.com>